+++ bar/bar/files.h
@@ -443,16 +443,16 @@ const char *File_getSystemTmpDirectory(void);
                                 const String directory
                                );
 #else /* not NDEBUG */
-  Errors __File_getTmpFile(const char    *__fileName__,
-                           ulong        __lineNb__,
+  Errors __File_getTmpFile(const char   *__fileName__,
+                           uint         __lineNb__,
                            FileHandle   *fileHandle,
                            const String pattern,
                            const String directory
                           );
-  Errors __File_getTmpFileCString(const char *__fileName__,
-                                  ulong      __lineNb__,
-                                  FileHandle  *fileHandle,
-                                  char const  *pattern,
+  Errors __File_getTmpFileCString(const char   *__fileName__,
+                                  uint         __lineNb__,
+                                  FileHandle   *fileHandle,
+                                  char const   *pattern,
                                   const String directory
                                  );
 #endif /* NDEBUG */
@@ -510,13 +510,13 @@ Errors File_openCString(FileHandle *fileHandle,
                        );
 #else /* not NDEBUG */
 Errors __File_open(const char   *__fileName__,
-                   ulong        __lineNb__,
+                   uint         __lineNb__,
                    FileHandle   *fileHandle,
                    const String fileName,
                    FileModes    fileMode
                   );
 Errors __File_openCString(const char *__fileName__,
-                          ulong      __lineNb__,
+                          uint       __lineNb__,
                           FileHandle *fileHandle,
                           const char *fileName,
                           FileModes  fileMode
@@ -541,7 +541,7 @@ Errors File_openDescriptor(FileHandle *fileHandle,
                           );
 #else /* not NDEBUG */
 Errors __File_openDescriptor(const char *__fileName__,
-                             ulong      __lineNb__,
+                             uint       __lineNb__,
                              FileHandle *fileHandle,
                              int        fileDescriptor,
                              FileModes  fileMode
@@ -560,7 +560,10 @@ Errors __File_openDescriptor(const char *__fileName__,
 #ifdef NDEBUG
 Errors File_close(FileHandle *fileHandle);
 #else /* not NDEBUG */
-Errors __File_close(const char *__fileName__, ulong __lineNb__, FileHandle *fileHandle);
+Errors __File_close(const char *__fileName__,
+                    uint       __lineNb__,
+                    FileHandle *fileHandle
+                   );
 #endif /* NDEBUG */
 
 /***********************************************************************\
